+++ src/Model/Workflow/WorkflowModel.php
@@ -22,9 +22,11 @@ use Contao\Model;
  *
  * @package Netzmacht\ContaoWorkflowBundle\Contao\Model
  *
- * @property int    $id      The workflow id.
- * @property array  $process The process definition.
- * @property string $type    The workflow type.
+ * @property int        $id           The workflow id.
+ * @property array      $process      The process definition.
+ * @property string     $type         The workflow type.
+ * @property string     $providerName Provider name.
+ * @property string     $label        Label.
  */
 final class WorkflowModel extends Model
 {